refactor: remove unnecesary duplication check
redesign of the CsvRepo and Collect for better separation of
responsabilities, now can add many duplicate inputs buth the method
getAll() of the collect aggregate not return duplicate inputs.

This simplify the source code and reduce the test necessary to write and
mantain, respecting the principle of non-duplication of the Collect aggregate.