refactor: i assign correct types of methods
I kept the model methods (Input) within the domain layer,
eliminated redundant dtos in the infrastructure layer and refactored a few tests.

In ddd the principle that must be respected is to keep everything in its own layer.
The DTO's always goes in the application layer, and never within the domain.